pyproject.toml
[tool.poetry]
name = "django_filter_groups"
version = "0.1.0"
description = ""
authors = ["pnoryk <pnoryk@mail.ru>"]
repository = "https://github.com/PNoryk/django-filter-groups"
packages = [
    { include = "django_filter_groups" }
]
include = ["readme.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.6"
Django = ">=2"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
commitizen = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
